+++ examples/main/src/App.jsx
@@ -1,69 +1,60 @@
 /* eslint-disable */
 import { hot } from 'react-hot-loader/root';
-import React, { Suspense } from 'react';
-import MobileDetect from 'mobile-detect';
-import { convertFromRaw, convertToRaw, EditorState } from '@wix/draft-js';
-import { normalizeInitialState } from 'wix-rich-content-common';
-import './App.css';
-import RichContentRawDataEditor from './RichContentRawDataEditor';
+import React, { PureComponent } from 'react';
+import { ReflexContainer, ReflexElement, ReflexSplitter } from 'react-reflex';
+import compact from 'lodash/compact';
+import flatMap from 'lodash/flatMap';
+import { convertToRaw, EditorState } from '@wix/draft-js';
+import {
+  ContentStateEditor,
+  ErrorBoundary,
+  Fab,
+  SectionHeader,
+  SectionContent,
+} from './Components';
+import {
+  generateKey,
+  getStateFromObject,
+  isMobile,
+  loadStateFromStorage,
+  saveStateToStorage,
+} from './utils';
 const Editor = React.lazy(() => import('./editor/Editor'));
 const Viewer = React.lazy(() => import('./viewer/Viewer'));
-import Resizable from 're-resizable';
-import startCase from 'lodash/startCase';
-import local from 'local-storage';
 
-// const whyDidYouRender = require('@welldone-software/why-did-you-render/dist/no-classes-transpile/umd/whyDidYouRender.min.js');
-// whyDidYouRender(React, { include: [/.*/], exclude: [/Decor|JSONI/] });
-
-const anchorTarget = '_top';
-const relValue = 'noreferrer';
-function Checkbox({ name, onChange, ...inputProps }) {
-  const handleChange = e => onChange(name, e.target.checked);
-  return (
-    <label>
-      <input type="checkbox" onChange={handleChange} {...inputProps} />
-      {startCase(name)}
-    </label>
-  );
-}
-
-const checkBoxes = [
-  'editor',
-  'viewer',
-  'contentStateEditor',
-  'mounted',
-  'staticToolbar',
-  'mobile',
-  'readOnly',
-];
-const mobileCheckBoxes = ['editor', 'viewer'];
-
-class App extends React.PureComponent {
+class App extends PureComponent {
   constructor(props) {
     super(props);
-    this.state = {
-      lastSave: new Date(),
-      editorState: EditorState.createEmpty(),
-      mounted: true,
-      contentStateEditor: true,
-      showDevToggles: true,
-      editor: true,
-      viewer: true,
-      ...this.getLocalStorageState(),
-    };
-    this.md = window ? new MobileDetect(window.navigator.userAgent) : null;
+    this.isMobile = isMobile();
+    this.state = this.getInitialState();
   }
 
-  getLocalStorageState() {
-    const state = {};
-    [...checkBoxes, 'contentWidth'].forEach(key => {
-      let val = local.get(key);
-      if (val !== null) {
-        state[key] = val;
-      }
-    });
-    return state;
+  getInitialState() {
+    const containerKey = generateKey('container');
+    const editorState = EditorState.createEmpty();
+    const localState = loadStateFromStorage();
+    if (localState) {
+      return { containerKey, editorState, ...JSON.parse(localState) };
+    } else {
+      return {
+        containerKey,
+        editorState,
+        isEditorShown: true,
+        isViewerShown: !this.isMobile,
+        isContentStateShown: false,
+      };
+    }
   }
+
+  componentDidMount() {
+    window && window.addEventListener('resize', this.onContentStateEditorResize);
+  }
+
+  componentWillUnmount() {
+    window && window.removeEventListener('resize', this.onContentStateEditorResize);
+  }
+
+  setContentStateEditor = ref => (this.contentStateEditor = ref);
 
   setViewerState = editorState => {
     const content = editorState.getCurrentContent();
@@ -72,157 +63,136 @@ class App extends React.PureComponent {
     }
   };
 
+  onContentStateEditorChange = obj => {
+    this.setState(getStateFromObject(obj));
+  };
+
+  onContentStateEditorResize = () =>
+    this.contentStateEditor && this.contentStateEditor.refreshLayout();
+
   onEditorChange = editorState => {
     const state = {
-      lastSave: new Date(),
       editorState,
     };
     this.setState(state);
     this.setViewerState(editorState);
   };
 
-  isMobileDevice = () => {
-    return this.md && this.md.mobile() !== null;
+  onSectionVisibilityChange = (sectionName, isVisible) => {
+    this.setState(
+      { [`is${sectionName}Shown`]: isVisible, containerKey: generateKey('prefix') },
+      () => {
+        saveStateToStorage(this.state);
+      }
+    );
+    this.onContentStateEditorResize();
   };
 
-  onRichContentRawDataEditorChange = obj => {
-    this.setState(this.convertJsObject(obj));
+  renderEditor = () => {
+    const { isEditorShown, editorState, staticToolbar } = this.state;
+    const settings = [];
+    if (!isMobile()) {
+      settings.push({
+        name: 'Static Toolbar',
+        action: () => this.setState(state => ({ staticToolbar: !state.staticToolbar })),
+      });
+    }
+    return (
+      isEditorShown && (
+        <ReflexElement key="editor-section" className="section">
+          <SectionHeader
+            title="Editor"
+            settings={settings}
+            onHide={this.onSectionVisibilityChange}
+          />
+          <SectionContent>
+            <ErrorBoundary>
+              <Editor
+                onChange={this.onEditorChange}
+                editorState={editorState}
+                isMobile={this.isMobile}
+                staticToolbar={staticToolbar}
+              />
+            </ErrorBoundary>
+          </SectionContent>
+        </ReflexElement>
+      )
+    );
   };
 
-  convertJsObject(obj) {
-    const normalizedState = normalizeInitialState(obj, {
-      anchorTarget,
-      relValue,
-    });
-    const editorState = EditorState.createWithContent(convertFromRaw(normalizedState));
-    return { editorState, viewerState: normalizedState };
-  }
+  renderViewer = () => {
+    const { isViewerShown, viewerState } = this.state;
+    return (
+      isViewerShown && (
+        <ReflexElement key="viewer-section" className="section">
+          <SectionHeader title="Viewer" onHide={this.onSectionVisibilityChange} />
+          <SectionContent>
+            <ErrorBoundary>
+              <Viewer initialState={viewerState} isMobile={this.isMobile} />
+            </ErrorBoundary>
+          </SectionContent>
+        </ReflexElement>
+      )
+    );
+  };
 
-  onCheckBoxChange = (name, checked) => {
-    this.setState({ [name]: checked });
-    local.set(name, checked);
+  renderContentState = () => {
+    const { isContentStateShown, editorState } = this.state;
+    return (
+      isContentStateShown && (
+        <ReflexElement
+          key="contentstate-section"
+          className="section"
+          onStopResize={this.onContentStateEditorResize}
+        >
+          <SectionHeader title="Content State" onHide={this.onSectionVisibilityChange} />
+          <SectionContent isLoadedLazily={false}>
+            <ContentStateEditor
+              ref={this.setContentStateEditor}
+              onChange={this.onContentStateEditorChange}
+              contentState={convertToRaw(editorState.getCurrentContent())}
+            />
+          </SectionContent>
+        </ReflexElement>
+      )
+    );
+  };
+
+  setSectionVisibility = (sectionName, isVisible) =>
+    this.setState({ [`show${sectionName}`]: isVisible });
+
+  renderSections = () => {
+    const sections = compact([this.renderEditor(), this.renderViewer(), this.renderContentState()]);
+
+    return flatMap(sections, (val, i, arr) =>
+      arr.length - 1 !== i
+        ? [val, <ReflexSplitter className="splitter" propagate={true} key={`splitter-${i}`} />]
+        : val
+    );
   };
 
   render() {
-    const {
-      showDevToggles,
-      mobile: simulateMobile,
-      editor: showEditor,
-      viewer: showViewer,
-      contentStateEditor: showContentStateEditor,
-    } = this.state;
-    const isMobileDevice = this.isMobileDevice();
-    const isMobile = simulateMobile || isMobileDevice;
-    const editor = (
-      <Suspense fallback={<div>Loading...</div>}>
-        <ErrorBoundary>
-          <Editor
-            onChange={this.onEditorChange}
-            editorState={this.state.editorState}
-            readOnly={this.state.readOnly}
-            isMobile={isMobile}
-            staticToolbar={this.state.staticToolbar}
-          />
-        </ErrorBoundary>
-      </Suspense>
-    );
-    const viewer = (
-      <Suspense fallback={<div>Loading...</div>}>
-        <ErrorBoundary>
-          <Viewer initialState={this.state.viewerState} mobile={isMobile} config={this.config} />
-        </ErrorBoundary>
-      </Suspense>
-    );
-    const checkBoxComponents = (isMobileDevice ? mobileCheckBoxes : checkBoxes).map((name, i) => (
-      <div className="toggle" key={i}>
-        <Checkbox name={name} checked={this.state[name]} onChange={this.onCheckBoxChange} />
-      </div>
-    ));
+    const { isEditorShown, isViewerShown, isContentStateShown } = this.state;
+    const showEmptyState = !isEditorShown && !isViewerShown && !isContentStateShown;
+
     return (
       <div className="wrapper">
-        <div className="container">
-          {!isMobileDevice && (
-            <div className="header">
-              <h1 onClick={() => this.setState({ showDevToggles: !showDevToggles })}>
-                Wix Rich Content Editor
-              </h1>
-              <div
-                className="toggle-container"
-                style={{ display: this.state.showDevToggles ? 'block' : 'none' }}
-              >
-                {checkBoxComponents}
-              </div>
-              <span className="intro">Last saved on {this.state.lastSave.toTimeString()}</span>
-            </div>
+        <ReflexContainer orientation="vertical" windowResizeAware={true} className="container">
+          {showEmptyState ? (
+            <div className="empty-state">Wix Rich Content</div>
+          ) : (
+            this.renderSections()
           )}
-          <div className="content">
-            {this.state.mounted && (
-              <div className="columns">
-                {isMobileDevice ? (
-                  <div className={'mobileDevice'} style={{ width: '100%' }}>
-                    <div style={{ display: 'flex' }}>{checkBoxComponents}</div>
-                    {showEditor && editor}
-                    {showViewer && viewer}
-                  </div>
-                ) : (
-                  <Resizable
-                    onResize={(event, direction, { clientWidth }) =>
-                      local.set('contentWidth', clientWidth)
-                    }
-                    defaultSize={{ width: this.state.contentWidth || '85%' }}
-                    className={'resizable'}
-                  >
-                    {showEditor && editor}
-                    {showViewer && viewer}
-                  </Resizable>
-                )}
-                {showContentStateEditor && !isMobileDevice && (
-                  <div className="column side">
-                    <RichContentRawDataEditor
-                      onChange={this.onRichContentRawDataEditorChange}
-                      content={convertToRaw(this.state.editorState.getCurrentContent())}
-                    />
-                  </div>
-                )}
-              </div>
-            )}
-          </div>
-        </div>
+        </ReflexContainer>
+        <Fab
+          isMobile={this.isMobile}
+          isEditorShown={isEditorShown}
+          isViewerShown={isViewerShown}
+          isContentStateShown={isContentStateShown}
+          toggleSectionVisibility={this.onSectionVisibilityChange}
+        />
       </div>
     );
-  }
-}
-
-class ErrorBoundary extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  componentDidCatch(error, info) {
-    this.setState({ error, info });
-  }
-
-  retry = () => {
-    this.setState({ error: undefined });
-  };
-
-  render() {
-    const { error, info } = this.state;
-    if (error) {
-      return (
-        <div style={{ whiteSpace: 'pre-wrap', marginLeft: '5px' }}>
-          <h1> oh oh :(</h1>
-          <button style={{ fontSize: '20px', background: 'aliceblue' }} onClick={this.retry}>
-            Retry
-          </button>
-          <p>Look for more info in the console</p>
-          <p>{error + info.componentStack}</p>
-        </div>
-      );
-    }
-
-    return this.props.children;
   }
 }