+++ examples/main/src/index.js
@@ -2,7 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
-import './index.css';
+import 'react-reflex/styles.css';
+import './styles.global.scss';
 
 ReactDOM.render(<App />, document.getElementById('root'));
 registerServiceWorker();