+++ packages/plugin-text-color/src/index.js
@@ -1,4 +1,4 @@
 export { createTextColorPlugin } from './createTextColorPlugin';
 export { TEXT_COLOR_TYPE } from './types';
 export { Modals, ModalsMap } from './modals';
-export { default as createTextColorDecorator } from './text-color-decorator';
+export { default as textColorInlineStyleMapper } from './textColorInlineStyleMapper';