+++ packages/viewer/src/utils/convertContentState.jsx
@@ -7,6 +7,7 @@ import endsWith from 'lodash/endsWith';
 import List from '../List';
 import getPluginsViewer from '../PluginsViewer';
 import { getTextDirection } from './textUtils';
+import { staticInlineStyleMapper } from '../staticInlineStyleMapper';
 
 const isEmptyContentState = raw =>
   !raw || !raw.blocks || (raw.blocks.length === 1 && raw.blocks[0].text === '');
@@ -30,18 +31,8 @@ const kebabToCamelObjectKeys = (obj = {}) =>
 
 const blockDataToStyle = ({ dynamicStyles }) => kebabToCamelObjectKeys(dynamicStyles);
 
-const getInline = mergedStyles => {
-  return {
-    BOLD: (children, { key }) => <strong key={key}>{children}</strong>,
-    ITALIC: (children, { key }) => <em key={key}>{children}</em>,
-    UNDERLINE: (children, { key }) => <u key={key}>{children}</u>,
-    CODE: (children, { key }) => (
-      <span key={key} className={mergedStyles.code}>
-        {children}
-      </span>
-    ),
-  };
-};
+const getInline = (inlineStyleMappers, mergedStyles) =>
+  combineMappers([...inlineStyleMappers, staticInlineStyleMapper], mergedStyles);
 
 const getBlocks = (mergedStyles, textDirection) => {
   const getList = ordered => (items, blockProps) => {
@@ -123,12 +114,12 @@ const normalizeContentState = contentState => ({
   }),
 });
 
-const combineTypeMappers = mappers => {
+const combineMappers = (mappers, ...args) => {
   if (!mappers || !mappers.length || mappers.some(resolver => typeof resolver !== 'function')) {
-    console.warn('typeMappers is expected to be a function array'); // eslint-disable-line no-console
+    console.warn(`${mappers} is expected to be a function array`); // eslint-disable-line no-console
     return {};
   }
-  return mappers.reduce((map, mapper) => Object.assign(map, mapper()), {});
+  return mappers.reduce((map, mapper) => Object.assign(map, mapper(...args)), {});
 };
 
 const redraftOptions = {
@@ -153,6 +144,7 @@ const convertToReact = (
   typeMap,
   entityProps,
   decorators,
+  inlineStyleMappers,
   options = {}
 ) => {
   if (isEmptyContentState(contentState)) {
@@ -162,9 +154,9 @@ const convertToReact = (
   return redraft(
     normalizeContentState(contentState),
     {
-      inline: getInline(mergedStyles),
+      inline: getInline(inlineStyleMappers, mergedStyles),
       blocks: getBlocks(mergedStyles, textDirection),
-      entities: getEntities(combineTypeMappers(typeMap), entityProps, mergedStyles),
+      entities: getEntities(combineMappers(typeMap), entityProps, mergedStyles),
       decorators,
     },
     { ...redraftOptions, ...options }